+++ cratedb_toolkit/iac/aws.py
@@ -1,9 +1,10 @@
+from lorrystream.carabas.aws import DynamoDBKinesisPipe, RDSPostgreSQLDMSKinesisPipe
 from lorrystream.carabas.aws.function.model import LambdaFactory
 from lorrystream.carabas.aws.function.oci import LambdaPythonImage
-from lorrystream.carabas.aws.stack import DynamoDBKinesisPipe
 
 __all__ = [
+    "DynamoDBKinesisPipe",
     "LambdaFactory",
     "LambdaPythonImage",
-    "DynamoDBKinesisPipe",
+    "RDSPostgreSQLDMSKinesisPipe",
 ]